Move data insertion helpers to tests/helpers.py
